+++ python/nautilus_trader/analysis/config.py
@@ -77,9 +77,14 @@ class TearsheetDrawdownChart(TearsheetChart):
 
 @dataclass(frozen=True, kw_only=True)
 class TearsheetMonthlyReturnsChart(TearsheetChart):
+    compounding: bool = True
+
     @property
     def name(self) -> str:
         return "monthly_returns"
+
+    def kwargs(self) -> dict[str, Any]:
+        return {"compounding": self.compounding}
 
 
 @dataclass(frozen=True, kw_only=True)
@@ -98,9 +103,14 @@ class TearsheetRollingSharpeChart(TearsheetChart):
 
 @dataclass(frozen=True, kw_only=True)
 class TearsheetYearlyReturnsChart(TearsheetChart):
+    compounding: bool = True
+
     @property
     def name(self) -> str:
         return "yearly_returns"
+
+    def kwargs(self) -> dict[str, Any]:
+        return {"compounding": self.compounding}
 
 
 @dataclass(frozen=True, kw_only=True)